requirements.txt
# Core dependencies
streamlit>=1.24.0
tensorflow>=2.13.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.8.0

# TensorFlow Lite (included with tensorflow for Python 3.13)
tensorflow>=2.15.0

# Web framework (already listed above)

# Utilities
pandas>=2.0.0
pathlib>=1.0.0

# Optional dependencies for enhanced functionality
# Uncomment if needed:
# streamlit-aggrid>=0.3.0
# plotly>=5.15.0
# streamlit-lottie>=0.0.11